backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.0
pydantic==2.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
